+++ python_modules/dagster-graphql/dagster_graphql/implementation/context.py
@@ -2,12 +2,16 @@ from dagster import ExecutionTargetHandle, check
 from dagster.core.instance import DagsterFeatures, DagsterInstance
 
 from .pipeline_execution_manager import PipelineExecutionManager
+from .reloader import Reloader
 
 
 class DagsterGraphQLContext(object):
-    def __init__(self, handle, execution_manager, instance, raise_on_error=False, version=None):
+    def __init__(
+        self, handle, execution_manager, instance, reloader=None, raise_on_error=False, version=None
+    ):
         self._handle = check.inst_param(handle, 'handle', ExecutionTargetHandle)
         self.instance = check.inst_param(instance, 'instance', DagsterInstance)
+        self.reloader = check.opt_inst_param(reloader, 'reloader', Reloader)
         self.execution_manager = check.inst_param(
             execution_manager, 'pipeline_execution_manager', PipelineExecutionManager
         )