+++ CHANGES.md
@@ -1,6 +1,12 @@
 # Changelog
 
+## Upcoming (0.6.0)
+
+- A "Reload" button in the top right in dagit restarts the web-server process and updates
+  the UI to reflect repo changes, including DAG structure, solid names, type names, etc.
+
 ## 0.5.9
+
 - Fixes an issue using custom types for fan-in dependencies with intermediate storage.
 
 ## 0.5.8